+++ app/controllers/agreements_controller.rb
@@ -1,5 +1,8 @@
 class AgreementsController < ApplicationController
   def index
-    @agreements = API::Agreement.includes(:framework, :supplier).all
+    @agreements = API::Agreement.includes(:framework, :supplier)
+                                .where(active: params[:status])
+                                .all
+                                .sort_by! { |t| t.framework.name }
   end
 end