+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [release-84] - 2022-10-??
+
+- RMI-33: New agreements page
+- RMI-419: Added status filter to Agreements page 
+
 ## [release-83] - 2022-09-29
 
 - RMI-509: Amended wording to suppliers upload and ingest failed pages
@@ -448,6 +453,7 @@
 
 Initial release
 
+[release-84]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-83...release-84
 [release-83]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-82...release-83
 [release-82]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-81...release-82
 [release-81]: https://github.com/Crown-Commercial-Service/DataSubmissionService/compare/release-80...release-81